test accurary before recalibration - step 3,4

+++ app/src/main/java/com/calorieko/app/viewmodel/LogMealViewModel.kt
@@ -43,8 +43,8 @@ class LogMealViewModel(
     private val _phase = MutableStateFlow(LogMealPhase.SCANNING)
     val phase: StateFlow<LogMealPhase> = _phase.asStateFlow()
 
-    private val _weight = MutableStateFlow(0)
-    val weight: StateFlow<Int> = _weight.asStateFlow()
+    private val _weight = MutableStateFlow(0f)
+    val weight: StateFlow<Float> = _weight.asStateFlow()
 
     private val _weightStable = MutableStateFlow(false)
     val weightStable: StateFlow<Boolean> = _weightStable.asStateFlow()
@@ -105,7 +105,7 @@ class LogMealViewModel(
         }
     }
 
-    fun updateRealWeight(realWeight: Int) {
+    fun updateRealWeight(realWeight: Float) {
         _weight.value = realWeight
         _weightStable.value = false
         weightStabilizationJob?.cancel()
@@ -142,7 +142,7 @@ class LogMealViewModel(
     fun processCapture() {
         val results = _latestResults.value
         val currentWeight = _weight.value
-        if (results.isEmpty() || !_weightStable.value || currentWeight <= 0) return
+        if (results.isEmpty() || !_weightStable.value || currentWeight <= 0f) return
 
         val top1 = results.getOrNull(0)
         val top2 = results.getOrNull(1)
@@ -213,7 +213,7 @@ class LogMealViewModel(
         viewModelScope.launch {
             val food = withContext(Dispatchers.IO) { foodDao.getFoodByName(dishName) }
             if (food != null) {
-                val w = currentWeight.toFloat()
+                val w = currentWeight
                 val dish = LoggedDish(
                     dishNameEn = dishName,
                     weightGrams = w,
